# Salus/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "medical API",
        "description": "API for medical advices and finding doctors.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/doctors": {
            "get": {
                "tags": [
                    "doctors"
                ],
                "summary": "Get all doctors",
                "description": "Display a listing of the resource.",
                "operationId": "cd066ac893c261ef13773c81ebc8c53e",
                "responses": {
                    "200": {
                        "description": "List of doctors"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "doctors not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/doctors/{id}": {
            "get": {
                "tags": [
                    "doctors"
                ],
                "summary": "Get all doctors",
                "description": "Display the specified resource.",
                "operationId": "f7d27b7312d2bcfb24da4bf9b7109e21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "fetch doctor"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "doctors not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/doctors/search": {
            "get": {
                "tags": [
                    "doctors"
                ],
                "summary": "search for doctors",
                "operationId": "e08889bdf474075b5e46a6efd9f43f9b",
                "parameters": [
                    {
                        "name": "specialty",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "General Medicine"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "rabat"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of doctors"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "doctors not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}